Distance/threshold test for toolmarks

Compute all possible correlations for windows of length n between the class components. Determine the location of the maximized correlation. Given this location, create a diamond around it in the individual matrix of correlations For each offset in this diamond, compute the maximized correlation (1) Determine the distance between the offset for the class and indiviudal components (2) Compute the Threshold test statistics

fixed_width_no_modeling(dat1, dat2, coarse = 0.25, fine = 0.01,
  window.size = 0.6, M = 500)

Arguments

dat1
a one column matrix representing a digitized tool mark

dat2
a one column matrix representing a second digitized tool mark

coarse
normalization smoothing parameter

fine
decomposition smoothing parameter

window.size
desired window size for the correlations to compute

M
search area restriction

Value

list with

"max_corr"maximized indiviudal component correlation

"Smooth_offset"optimal Class offset

"Resid_offset"optimal individual offset

"dist_pval" distance p-value

"thresh_pval" threshold p-value

"Above" Number of offsets with correlation bigger than threshold

"total_thresh" 2*M+1

"mark1_decompostion" plot of decomposition d1

"mark2_decompostion" plot of decomposition d2

"class_correlations" plot of class correlation

"individual_correlations" plot of individual correlation

"distance_plot" distance_plot

"threshold_plot" threshold_plot